use return types
