--- src/Illuminate/Database/Eloquent/Casts/AsAttribute.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Illuminate\Database\Eloquent\Casts;
-
-use Attribute as BaseAttribute;
-
-#[BaseAttribute]
-class AsAttribute
-{
-    //
-}
